refactor: create api sub package
