+++ compliance_test.go
@@ -7,6 +7,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/jmespath-community/go-jmespath/pkg/api"
 	"github.com/jmespath-community/go-jmespath/pkg/parsing"
 	"github.com/stretchr/testify/assert"
 )
@@ -103,7 +104,7 @@ func runSyntaxTestCase(assert *assert.Assertions, given interface{}, testcase Te
 	// Anything with an .Error means that we expect that JMESPath should return
 	// an error when we try to evaluate the expression.
 	//fmt.Println(fmt.Sprintf("%s: %s", filename, testcase.Expression))
-	_, err := Search(testcase.Expression, given)
+	_, err := api.Search(testcase.Expression, given)
 	assert.NotNil(err, fmt.Sprintf("Expression: %s", testcase.Expression))
 }
 
@@ -123,7 +124,7 @@ func runTestCase(assert *assert.Assertions, given interface{}, testcase TestCase
 		assert.Fail(errMsg)
 		return
 	}
-	actual, err := Search(testcase.Expression, given)
+	actual, err := api.Search(testcase.Expression, given)
 	if assert.Nil(err, fmt.Sprintf("Expression: %s", testcase.Expression)) {
 		assert.Equal(testcase.Result, actual, fmt.Sprintf("Expression: %s", testcase.Expression))
 	}
