+++ pkg/api/api.go
@@ -1,4 +1,4 @@
-package jmespath
+package api
 
 import (
 	"strconv"
